Move lookup data source Read() into datasource.go
Consolidate Read() and its state-mapping helpers from read.go into
datasource.go for the backup_job, cluster_layout, monitoring_check_type,
network_pool_server_type, network_router_type and subnet_type data
sources, matching the convention used by the provider's other data
sources. Pure reorganization; no schema or behavior change.
